\name{reflect_315}
\alias{reflect_315}
\title{Reflect Matrix at 315 degrees}
\description{
  Reflect a matrix at 315 degrees.
}
\usage{
  reflect_315(x)
}
\arguments{
  \item{x}{A matrix to be reflected at 315 degrees.}
}
\details{
  This function reflects a matrix \code{x} at 315 degrees, applying vertical reflection, horizontal reflection, and diagonal reflection at 45 degrees in sequence.
}
\value{
  A matrix resulting from the 315-degree reflection of \code{x}.
}
\author{
  Bret Beheim \\ \email{bret_beheim@eva.mpg.de}
}
\examples{
  matrix1 <- matrix(1:9, ncol = 3)
  result_matrix <- reflect_315(matrix1)
  print(result_matrix)
}
